requirements.txt
APScheduler==3.10.4
beautifulsoup4==4.12.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.36
openpyxl==3.1.5